feat(discover): Remove get_filter
- removes the get filter function since its been replaced by the QueryBuilder
